Move state explanation components into the top-level namespace
This will allow us to re-use these components in the API docs.
