+++ app/components/state_event_explanation_component.html.erb
