+++ app/views/support_interface/docs/candidate_flow.html.erb
@@ -23,6 +23,6 @@
 </details>
 
 <% ProcessState.workflow_spec.states.each do |_, state| %>
-  <%= render SupportInterface::StateExplanationComponent, state: state, machine: ProcessState %>
+  <%= render StateExplanationComponent, state: state, machine: ProcessState %>
   <hr class='govuk-section-break govuk-section-break--xl govuk-section-break--visible'>
 <% end %>
